Add Map registration (MapIDs) + Overall map play-time tracking
Map IDs are one digit ints which are in the order they are created and
are not server specific. If you wanted to get the total time a map was
played on all servers, you can select * from map_playtimes where mid =
'%s' and add the results.

These can be accessed while Sys_InMap() is true from Sys_GetMapID(),
assuming it has already loaded. You can use the forward OnMapIDLoaded to
get the ID as soon as it is loaded.